feat: Se actualizó account service

+++ account-service/src/main/java/com/novabank/account/exception/CustomerNotFoundException.java
@@ -1,9 +1,9 @@
 package com.novabank.account.exception;
 
 public class CustomerNotFoundException extends RuntimeException {
-    private final Integer customer_id;
+    private final Long customer_id;
 
-    public CustomerNotFoundException(Integer id) {
+    public CustomerNotFoundException(Long id) {
         super("Cliente no encontrado con ID: " + id);
         this.customer_id = id;
     }
@@ -13,7 +13,7 @@ public class CustomerNotFoundException extends RuntimeException {
         this.customer_id = null;
     }
 
-    public Integer getCustomer_id() {
+    public Long getCustomer_id() {
         return customer_id;
     }
 }
